requirements.txt
flask
tensorflow
numpy
flask-cors
python-dotenv
pillow